+++ internal/commands/router/create.go
@@ -6,13 +6,16 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/ui"
+
+	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud/request"
 	"github.com/spf13/pflag"
 )
 
 type createCommand struct {
 	*commands.BaseCommand
-	name string
+	name         string
+	staticRoutes []string
 }
 
 // CreateCommand creates the "router create" command
@@ -22,7 +25,7 @@ func CreateCommand() commands.Command {
 			"create",
 			"Create a router",
 			"upctl router create --name my_router",
-			`upctl router create --name "My Router"`,
+			`upctl router create --name "My Router" --static-route name=my_static_route,nexthop=10.0.0.100,route=0.0.0.0/0"`,
 		),
 	}
 }
@@ -31,6 +34,15 @@ func CreateCommand() commands.Command {
 func (s *createCommand) InitCommand() {
 	fs := &pflag.FlagSet{}
 	fs.StringVar(&s.name, "name", s.name, "Router name.")
+	fs.StringArrayVar(
+		&s.staticRoutes,
+		"static-route",
+		[]string{},
+		"Static route for this router, multiple can be declared.\n\n "+
+			"Fields: \n"+
+			"  name: string \n"+
+			"  nexthop: string \n"+
+			"  route: string")
 
 	s.AddFlags(fs)
 	_ = s.Cobra().MarkFlagRequired("name")
@@ -46,7 +58,20 @@ func (s *createCommand) ExecuteWithoutArguments(exec commands.Executor) (output.
 	msg := fmt.Sprintf("Creating router %s", s.name)
 	exec.PushProgressStarted(msg)
 
-	res, err := exec.Network().CreateRouter(exec.Context(), &request.CreateRouterRequest{Name: s.name})
+	var staticRoutes []upcloud.StaticRoute
+	for _, v := range s.staticRoutes {
+		staticRoute, err := handleStaticRoute(v)
+		if err != nil {
+			return commands.HandleError(exec, msg, err)
+		}
+
+		staticRoutes = append(staticRoutes, staticRoute)
+	}
+
+	res, err := exec.Network().CreateRouter(exec.Context(), &request.CreateRouterRequest{
+		Name:         s.name,
+		StaticRoutes: staticRoutes,
+	})
 	if err != nil {
 		return commands.HandleError(exec, msg, err)
 	}